Planet Generatio #2

+++ BLL/BLL/Utilities/RandomNumbers.cs
@@ -22,7 +22,7 @@ namespace BLL.Utilities
         /// <param name="max"></param>
         /// <param name="rnd"></param>
         /// <returns></returns>
-        public static double RandomDouble(int min, int max, Random rnd)
+        public static double RandomDouble(double min, double max, Random rnd)
         {
             return rnd.NextDouble() * (max - min) + min;
         }
